警告を修正

+++ iBeaconSampleCentral/iBeaconSampleCentral/ViewController.m
@@ -109,8 +109,8 @@
                 break;
         }
         
-        NSString *message = [NSString stringWithFormat:@"major:%@, minor:%@, accuracy:%f, rssi:%d",
-                             nearestBeacon.major, nearestBeacon.minor, nearestBeacon.accuracy, nearestBeacon.rssi];
+        NSString *message = [NSString stringWithFormat:@"major:%@, minor:%@, accuracy:%f, rssi:%ld",
+                             nearestBeacon.major, nearestBeacon.minor, nearestBeacon.accuracy, (long)nearestBeacon.rssi];
         [self sendLocalNotificationForMessage:[rangeMessage stringByAppendingString:message]];
     }
 }
